#21: Refactor buildApiReq (tested)
