+++ pkg/api/buildApiReq.go
@@ -21,7 +21,6 @@ import (
 	"net/http"
 
 	"github.com/cloudwego/hertz/pkg/app"
-	"github.com/go-playground/validator/v10"
 )
 
 const provider = "google" // placeholder for testing until provider pool is implemented
@@ -64,15 +63,9 @@ func sendRequest(payload []byte) (interface{}, error) {
 	}
 
 	// Create the full URL
-	var url string
-	switch provider {
-	case "google":
-		url = requestDetails.ApiConfig.BaseURL
-	default:
-		url = requestDetails.ApiConfig.BaseURL + requestDetails.ApiConfig.Endpoint
-	}
+	url := buildProviderURL(provider, requestDetails.ApiConfig)
 
-	slog.Info("rovider URL: " + url)
+	slog.Debug("provider URL: " + url)
 
 	// Marshal the requestDetails.RequestBody struct into JSON
 	body, err := json.Marshal(requestDetails.RequestBody)
@@ -81,152 +74,78 @@ func sendRequest(payload []byte) (interface{}, error) {
 		return nil, err
 	}
 
-	slog.Info("request Body: " + string(body))
+	slog.Debug("request Body: " + string(body))
 
 	// Create a new request using http
 	req, err := http.NewRequest("POST", url, bytes.NewBuffer(body))
-
-	// If there was an error with creating the request, handle it
 	if err != nil {
 		slog.Error("Error creating request: %v", err)
 		return nil, err
 	}
 
 	// Set the headers
-	for key, values := range requestDetails.ApiConfig.Headers {
-		for _, value := range values {
-			req.Header.Set(key, value)
-		}
-	}
+	setRequestHeaders(req, requestDetails.ApiConfig.Headers)
 
 	// Send the request using http Client
 	client := &http.Client{}
-
 	resp, err := client.Do(req)
-
 	if err != nil {
 		slog.Error("Error sending request: ", err)
+		return nil, err
+	}
+	defer resp.Body.Close()
+
+	responseBody, err := io.ReadAll(resp.Body)
+	if err != nil {
+		slog.Error("error reading response body: %v", err)
+		return nil, err
 	}
 
-	// retrieve payload from response
-	body, _ = io.ReadAll(resp.Body)
-
-	// Unmarshal the JSON response into a map
 	var responseMap map[string]interface{}
-	err = json.Unmarshal(body, &responseMap)
-	if err != nil {
-		slog.Error("Error decoding JSON response:", err)
+	if err := json.Unmarshal(responseBody, &responseMap); err != nil {
+		slog.Error("error decoding JSON response: %v", err)
+		return nil, err
 	}
 
 	return responseMap, nil
 
 }
 
+// definePayload takes the client payload and returns the request body for the provider as a struct.
 func definePayload(payload []byte, endpoint string) (providers.RequestDetails, string, error) {
-
-	// this function takes the client payload and returns the request body for the provider as a struct
-
 	slog.Info("definePayload Function Called")
 
-	// Define a map to hold the JSON data. Should these be declared at the top?
-	var defaultConfig interface{}                         // Assuming defaultConfig is a struct
-	var finalApiConfig providers.ProviderDefinedApiConfig // Assuming finalApiConfig is a struct
-	var payload_data map[string]interface{}
-	var params map[string]interface{}
-	var api_key string
-	var endpoints []map[string]interface{}
-	var defaultConfigMap map[string]interface{}
-	var paramsMap map[string]interface{}
-
-	// Parse the JSON
-	err := json.Unmarshal([]byte(payload), &payload_data)
-	if err != nil {
-		// Handle error
+	var payloadData map[string]interface{}
+	if err := json.Unmarshal(payload, &payloadData); err != nil {
 		slog.Error("Error unmarshalling payload: %v", err)
+		return providers.RequestDetails{}, "", err
 	}
 
-	jsonString, _ := json.Marshal(payload_data)
-
-	slog.Debug("Payload: " + string(jsonString))
-
-	endpointsInterface := payload_data["endpoints"].([]interface{})
-
-	for _, endpoint := range endpointsInterface {
-		endpointMap, ok := endpoint.(map[string]interface{})
-		if !ok {
-			slog.Error("Invalid endpoint format")
-			continue
-		}
-		endpoints = append(endpoints, endpointMap)
-	}
-
-	jsonString, _ = json.Marshal(endpoints)
-
-	slog.Debug("Endpoint: " + string(jsonString))
-
-	providerList := make([]string, len(endpoints))
-	for i, endpoint := range endpoints {
-		provider, ok := endpoint["provider"].(string)
-		if !ok {
-			// Handle error
-			fmt.Println("provider not found")
-		}
-
-		providerList[i] = provider
-	}
-
-	slog.Info("provider list: " + fmt.Sprintf("%v", providerList))
-
-	// TODO: Send the providerList to the provider pool to get the provider selection. Mode list can be used as well. Mode is the routing strategy.
-	//modeList := payload_data["mode"].([]interface{})
-	// provider is used in sendRequest function to build Google URL. This might need to change or provider should be a const
-
-	for _, endpoint := range endpoints {
-		if endpoint["provider"] == provider {
-			params, _ = endpoint["params"].(map[string]interface{})
-			api_key, _ = endpoint["api_key"].(string)
-			break
-		}
-	}
-
-	// retrieve the default configs for the provider
-	defaultConfig, finalApiConfig, _ = buildApiConfig(provider, api_key, endpoint)
-
-	// convert the defaultConfig and params to maps
-	defaultConfigJson, _ := json.Marshal(defaultConfig)
-
-	paramsJson, _ := json.Marshal(params)
-
-
-	json.Unmarshal(defaultConfigJson, &defaultConfigMap)
-
-	json.Unmarshal(paramsJson, &paramsMap)
-
-	for key, value := range paramsMap {
-		if _, ok := defaultConfigMap[key]; ok {
-			defaultConfigMap[key] = value
-		}
-	}
-
-	updatedConfigJson, _ := json.Marshal(defaultConfigMap)
-
-	err = json.Unmarshal(updatedConfigJson, &defaultConfig)
+	// This is where the provider 
+	provider, err := extractProvider(payloadData)
 	if err != nil {
-		slog.Error("error occurred during unmarshalling. %v", err)
+		return providers.RequestDetails{}, "", err
 	}
 
-	//err = validateStruct(defaultConfig)
+	params, apiKey, err := extractEndpointDetails(payloadData, provider)
+	if err != nil {
+		return providers.RequestDetails{}, "", err
+	}
 
-	//if err != nil {
-	//	slog.Error("error occurred during validation.", err)
-	//	return providers.RequestDetails{}, err
-	//}
+	defaultConfig, finalApiConfig, err := buildApiConfig(provider, apiKey, endpoint)
+	if err != nil {
+		return providers.RequestDetails{}, "", err
+	}
 
-	slog.Info("default Config: " + fmt.Sprintf("%v", defaultConfig))
+	mergedConfig, err := mergeConfigs(defaultConfig, params)
+	if err != nil {
+		return providers.RequestDetails{}, "", err
+	}
 
-	var requestDetails providers.RequestDetails = providers.RequestDetails{RequestBody: defaultConfig, ApiConfig: finalApiConfig}
-
-	slog.Debug("requestDetails: " + fmt.Sprintf("%v", requestDetails))
+	requestDetails := providers.RequestDetails{
+		RequestBody: mergedConfig,
+		ApiConfig:   finalApiConfig,
+	}
 
 	return requestDetails, provider, nil
 }
@@ -279,32 +198,104 @@ func buildApiConfig(provider string, api_key string, endpoint string) (interface
 	return defaultConfig, finalApiConfig, nil
 }
 
+// Helper functions below:
 
-func validateStruct(config interface{}) error {
-    // Validate a struct
-	validate := validator.New()
-    err := validate.Struct(config)
-    if err != nil {
-        if invalidErr, ok := err.(*validator.InvalidValidationError); ok {
-            return fmt.Errorf("invalid validation error: %v", invalidErr)
-        }
-
-        for _, err := range err.(validator.ValidationErrors) {
-            fmt.Println(err.Namespace())
-            fmt.Println(err.Field())
-            fmt.Println(err.StructNamespace())
-            fmt.Println(err.StructField())
-            fmt.Println(err.Tag())
-            fmt.Println(err.ActualTag())
-            fmt.Println(err.Kind())
-            fmt.Println(err.Type())
-            fmt.Println(err.Value())
-            fmt.Println(err.Param())
-            fmt.Println()
-        }
-
-        return fmt.Errorf("validation failed for the configuration")
-    }
-
-    return nil
+// buildProviderURL constructs the URL for the provider request.
+func buildProviderURL(provider string, apiConfig providers.ProviderDefinedApiConfig) string {
+	if provider == "google" {
+		return apiConfig.BaseURL
+	}
+	return apiConfig.BaseURL + apiConfig.Endpoint
 }
+
+// setRequestHeaders sets the headers for the HTTP request.
+func setRequestHeaders(req *http.Request, headers map[string][]string) {
+	for key, values := range headers {
+		for _, value := range values {
+			req.Header.Set(key, value)
+		}
+	}
+}
+
+// extractProvider extracts the provider from the payload data.
+func extractProvider(payloadData map[string]interface{}) (string, error) {
+	endpoints, ok := payloadData["endpoints"].([]interface{})
+	if !ok {
+		return "", errors.New("endpoints not found or invalid format")
+	}
+
+	for _, endpoint := range endpoints {
+		endpointMap, ok := endpoint.(map[string]interface{})
+		if !ok {
+			continue
+		}
+		if provider, ok := endpointMap["provider"].(string); ok {
+			return provider, nil
+		}
+	}
+	return "", errors.New("provider not found in endpoints")
+}
+
+// extractEndpointDetails extracts details like parameters and API key for a given provider.
+func extractEndpointDetails(payloadData map[string]interface{}, provider string) (map[string]interface{}, string, error) {
+	endpoints, ok := payloadData["endpoints"].([]interface{})
+	if !ok {
+		return nil, "", errors.New("endpoints not found or invalid format")
+	}
+
+	for _, endpoint := range endpoints {
+		endpointMap, ok := endpoint.(map[string]interface{})
+		if !ok || endpointMap["provider"] != provider {
+			continue
+		}
+		params, _ := endpointMap["params"].(map[string]interface{})
+		apiKey, _ := endpointMap["api_key"].(string)
+		return params, apiKey, nil
+	}
+	return nil, "", errors.New("endpoint details not found for provider")
+}
+
+// mergeConfigs merges the default configuration with the parameters provided.
+func mergeConfigs(defaultConfig interface{}, params map[string]interface{}) (interface{}, error) {
+	defaultConfigMap, err := structToMap(defaultConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	for key, value := range params {
+		defaultConfigMap[key] = value
+	}
+
+	return mapToStruct(defaultConfigMap, defaultConfig)
+}
+
+// structToMap converts a struct to a map.
+func structToMap(data interface{}) (map[string]interface{}, error) {
+	jsonData, err := json.Marshal(data)
+	if err != nil {
+		return nil, err
+	}
+
+	var resultMap map[string]interface{}
+	if err := json.Unmarshal(jsonData, &resultMap); err != nil {
+		return nil, err
+	}
+
+	return resultMap, nil
+}
+
+// mapToStruct converts a map to a struct based on the provided struct type.
+func mapToStruct(dataMap map[string]interface{}, structType interface{}) (interface{}, error) {
+	jsonData, err := json.Marshal(dataMap)
+	if err != nil {
+		return nil, err
+	}
+
+	resultStruct := structType
+	if err := json.Unmarshal(jsonData, &resultStruct); err != nil {
+		return nil, err
+	}
+
+	return resultStruct, nil
+}
+
